Move the shell lint to the style job and add an fs_time test
The lint only reads files, so it does not need a runner with the full
dependency set; the functional tests need a real bash 3.2 and stay on macOS.
